requirements.txt
pyyaml
sklearn
matplotlib
tensorboard
termcolor
regex
